parte_2/activity/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Automat==22.10.0
autopep8==2.3.1
beautifulsoup4==4.12.3
billiard==4.2.0
boto3==1.34.37
botocore==1.34.37
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
coverage==7.3.0
cryptography==41.0.7
cssselect==1.2.0
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.4
django-cors-headers==4.3.1
django-phonenumber-field==7.2.0
django-sendgrid-v5==0.9.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.6
et-xmlfile==1.1.0
filelock==3.13.1
future==1.0.0
googlemaps==4.5.3
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itemadapter==0.8.0
itemloaders==1.1.0
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
json5==0.9.14
kombu==5.3.5
lxml==5.1.0
MarkupSafe==2.1.5
numpy==1.26.2
openpyxl==3.1.2
packaging==23.1
pandas==2.1.4
parsel==1.8.1
phonenumbers==8.13.24
pillow==10.3.0
prompt-toolkit==3.0.43
Protego==0.3.0
proxyscrape==0.3.0
psycopg2==2.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.12.1
pycparser==2.21
PyDispatcher==2.0.7
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.1
python-http-client==3.3.7
pytz==2023.3
PyYAML==6.0.1
queuelib==1.6.2
redis==5.0.1
reportlab==4.2.2
requests==2.31.0
requests-file==1.5.1
s3transfer==0.10.0
Scrapy==2.11.2
sendgrid==6.11.0
service-identity==24.1.0
setuptools==70.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
stripe==8.5.0
tld==0.13
tldextract==5.1.1
Twisted==22.10.0
typing_extensions==4.9.0
tzdata==2023.3
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==1.26.7
user-agents==2.2.0
vine==5.1.0
w3lib==2.1.2
wcwidth==0.2.13
zope.interface==6.1